Ringback: Only play all videos once as call will have timed out
When the welcome video and the three 'you are being connected' type
videos have been played once, Riot web would time out if the user did
not answer. As such, the call will no longer be awaiting an answer and
the 'you are being connected' videos are no longer applicable.
Thereafter we just loop the 'welcome' video.