FullMeshConference: Replace existing with new call from same user
Sometimes when reloading the page or switching from the video conf to
the lobby and back, the call that existed before the switch/reload is
lingering, despite actually being dead as the RTCPeerConnection ICE
connection state takes a while to time out and reach the failed state.

If we hang up that call and instead answer the new, incoming call, we
end up with a functioning call instead of one person waiting for the
other to answer and the other wondering where the first person went.