app: Fix call and conference state management
It was made possible to not have a 1:1 call if no peerId was specified.
In this case we would have fallen through to the else case for every
room update which would then mean that the conference would never hang
up. I am really surprised the behaviour was not more egregious...